refactor: relocate schema bridge + share snapshot parser (MOB-3291)
Second code-review pass:
- Move toSqlDelightSchema() from SqlDelightSqkonDriver.kt to BridgeHelpers.kt,
  alongside the sibling wrapSqkonBinders / mapWithSqkonCursor adapters.
- Move SchemaMigrationTest's local extractSection() into SchemaSnapshotUtil
  as section(); the snapshot format now has a single owner for both writer
  and reader.
- Document the intentionally-retained SQLDelight plugin alias in
  build.gradle.kts (codegen unused, removed in Phase 7; harmless
  "no databases" warning until then).

Schema parity confirmed robust: SchemaSnapshotUtil.normalise() strips
comments and collapses whitespace, so DDL reformatting cannot break the
snapshot. jvmTest + compileDebugKotlinAndroid green.
